% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{med_dat}
\alias{med_dat}
\title{Mediation Example Dataset}
\format{
\subsection{\code{med_dat}}{

A list with three items:
\describe{
\item{Y}{numeric vector containing the outcome variable}
\item{M}{numeric matrix of 20 continuous mediators}
\item{A}{numeric vector containing the exposure variable}
...
}
}
}
\usage{
med_dat
}
\description{
A toy dataset for demonstrating mediation analysis in settings where there
are multiple mediators
}
\keyword{datasets}
