% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spcma.R
\name{spcma}
\alias{spcma}
\title{Sparse Principal Component Mediation Analysis for High-Dimensional Mediators}
\usage{
spcma(
  A,
  M,
  Y,
  adaptive = FALSE,
  var_per = 0.8,
  n_pc = NULL,
  fused = FALSE,
  gamma = 0,
  eps = 1e-04,
  maxsteps = 2000,
  per_jump = 0.7,
  boot = TRUE,
  boot_ci_type = "bca",
  sims = 1000,
  ci_level = 0.95
)
}
\arguments{
\item{A}{numeric vector containing exposure variable.}

\item{M}{numeric matrix of high-dimensional mediators.}

\item{Y}{numeric vector containing continuous outcome variable.}

\item{adaptive}{a logical variable for whether the number of PCs should be
determined based on the proportion of variance explained. Default is \code{TRUE}.}

\item{var_per}{a numeric variable with the designated proportion of variance
explained if \code{adaptive = TRUE}. Default is 0.8.}

\item{n_pc}{a numeric variable with the desired number of PCs for when
\code{adaptive = FALSE}. Default is \code{NULL}.}

\item{fused}{a logical variable for whether the fused LASSO should be used
instead of the ordinary LASSO. Default is \code{FALSE}.}

\item{gamma}{a numeric variable \code{>=0} indicating the ratio of the
ordinary LASSO penalty to the fusion penalty (see \code{\link[genlasso:fusedlasso]{genlasso::fusedlasso()}}).
Ignored if \code{fused = FALSE}. Default is 0.}

\item{eps}{numeric variable indicating the multiplier for the ridge penalty
in case \code{X} is rank deficient (see  \code{\link[genlasso]{genlasso}}).
Default is \code{1e-4}.}

\item{maxsteps}{an integer specifying the maximum number of steps for the
algorithm before termination (see \code{\link[genlasso]{genlasso}}). Default
is 2000.}

\item{per_jump}{a numeric value used for tuning parameter selection: the
quantile of cut-off for the total variance change under different tuning
parameter values. The higher the value, the sparser the loading matrix.
Default is \eqn{0.7}.}

\item{boot}{a logical value. If \code{FALSE} a quasi-Bayesian approximation
is used for confidence intervals; if \code{TRUE} nonparametric bootstrap will
be used (see \code{\link[mediation]{mediate}}). Default is \code{TRUE}.}

\item{boot_ci_type}{a character string indicating the type of bootstrap
confidence intervals for when \code{boot = TRUE}. If \code{"bca"},
bias-corrected and accelerated (BCa) confidence intervals will be estimated.
If \code{"perc"}, percentile confidence intervals will be estimated
(see \code{\link[mediation]{mediate}}). Default is "bca".}

\item{sims}{number of Monte Carlo draws for nonparametric bootstrap or
quasi-Bayesian approximation (see \code{\link[mediation]{mediate}}).
Default is 1000.}

\item{ci_level}{the designated confidence level. Default 0.95.}
}
\value{
A list containing:
\itemize{
\item{loadings: }{a matrix of the sparse PC loadings.}
\item{pcs: }{a matrix of the PCs.}
\item{var_explained: the cumulative proportion of variance explained by the PCs.}
\item{contributions: }{a containing the estimates, confidence intervals, and p-values of the mediation contributions}
\item{effects: }{a containing the estimates, confidence intervals, and p-values of the mediation contributions}
}
}
\description{
\code{spcma} applies sparse principal component mediation analysis to
mediation settings in which the mediators are high-dimensional.
}
\references{
Zhao, Y., Lindquist, M. A. & Caffo, B. S. Sparse principal
component based high-dimensional mediation analysis. Comput. Stat.
Data Anal. 142, 106835 (2020).
}
